+++ crates/krilla/src/object/color.rs
@@ -488,30 +488,4 @@ mod tests {
     fn color_space_srgb(sc: &mut SerializeContext) {
         sc.register_colorspace(ColorSpace::Srgb);
     }
-
-    #[snapshot(single_page, settings_18)]
-    fn icc_v2_srgb(page: &mut Page) {
-        let mut surface = page.surface();
-        surface.fill_path(&rect_to_path(50.0, 50.0, 100.0, 100.0), red_fill(1.0));
-    }
-
-    #[snapshot(single_page, settings_18)]
-    fn icc_v2_sgrey(page: &mut Page) {
-        let mut surface = page.surface();
-        surface.fill_path(&rect_to_path(50.0, 50.0, 100.0, 100.0), Fill::default());
-    }
-
-    #[visreg(all)]
-    fn cmyk_color(surface: &mut Surface) {
-        let path = rect_to_path(20.0, 20.0, 180.0, 180.0);
-
-        surface.fill_path(&path, cmyk_fill(1.0));
-    }
-
-    #[visreg(all, settings_6)]
-    fn cmyk_with_icc(surface: &mut Surface) {
-        let path = rect_to_path(20.0, 20.0, 180.0, 180.0);
-
-        surface.fill_path(&path, cmyk_fill(1.0));
-    }
 }
